+++ extensions/linkwarden/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Linkwarden Changelog
 
+## [Filter Links by Collection] - {PR_MERGE_DATE}
+
+- View the Date Modified and Collection of a link
+- Filter links by collection
+- Controlled validation with better error handling in `add`
+
 ## [Fix API Auth] - 2024-08-20
 
 - The API now uses `Bearer` authorization (fix #13613)